backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic-settings
bcrypt
python-jose[cryptography]
requests
sqlalchemy
